Python packaging
This allows installation using pip and the ability to import the
components as a library.
