+++ AWS CLI - Idaptive V1/idaptive/aws_cli/aws/__init__.py
