+++ AWS CLI - Idaptive V1/idaptive/aws_cli/aws/assumerolesaml.py
