+++ AWS CLI - Idaptive V1/idaptive/aws_cli/aws/util.py
