+++ AWS CLI - Idaptive V1/idaptive/aws_cli/config/apps.py
