+++ AWS CLI - Idaptive V1/idaptive/aws_cli/core/authresponse.py
